*Completed MatchListIndex Page; *Need to figure out MatchInfoPopupPage firebase datapull/async error

+++ VitruvianApp2017/External APIs/Firebase/EventMatchData.cs
@@ -6,7 +6,7 @@ namespace VitruvianApp2017
 		public string matchNumber { get; set; }
 		public int[] Blue { get; set; }
 		public int[] Red { get; set; }
-		public int matchTime { get; set; }
+		public long matchTime { get; set; }
 
 		// Pit Data
